requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.18.0
tree-sitter>=0.21.0
tree-sitter-python>=0.21.0
tree-sitter-javascript>=0.21.0
tree-sitter-java>=0.21.0

# Code analysis
gitpython>=3.1.40
pygments>=2.17.0
radon>=6.0.1

# Data processing
pandas>=2.1.0
numpy>=1.24.0
pydantic>=2.5.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
rich>=13.7.0
pyyaml>=6.0.1

# Optional: For model fine-tuning
# transformers>=4.36.0
# torch>=2.1.0
# datasets>=2.16.0
# peft>=0.7.0
